+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* HTTP Write uses polymorphic buffers.
 * WebSocket uses polymorphic buffers.
 * Add polymorphic buffer sequence types.
 * CML permits finding Boost outside the boost build tree.